[Dist/Debian/CMake] Enable VPU and MYRIAD options
This patch enables VPU and MYRIAD build options to use Intel Movidius
Neural Compute Stick 1 and 2. Currently, all of the plugins are packaged
into one, but later, they would be separated into sub-packages. Note
that the firmwares for the MVNCS 1 and 2 are excluded and they would be
provided other packages.

Change-Id: Ie7ad5fd2b4a50a3710acd957f00aefab62454863
Signed-off-by: Wook Song <wook16.song@samsung.com>