[Dist/Debian] Remove version requirement for libtbb-dev
The version strings used in openvino and libtbb do not follow the debian
standard. Therefore, since it is not possible to resolve build
dependency on libtbb-dev because of the invalid version string, the
openvino packages could not be built on the launchpad ppa. To work-aroud
this problem, this patch removes the version requirement for libtbb-dev
in the control file.

Change-Id: Ic4167c3b7b201fd02387975dda6781d1e48c581f
Signed-off-by: Wook Song <wook16.song@samsung.com>